refactoring: extract Boss Client, models & token repo in extra files
